requirements.txt
python-telegram-bot==13.15
wikipedia==1.4.0